fix dto got wrong value

+++ back-end/src/main/java/com/example/application/repository/SubCategoryRepository.java
@@ -3,5 +3,5 @@ package com.example.application.repository;
 import com.example.application.entity.SubCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubCategoryRepository extends JpaRepository<SubCategory, Integer> {
+public interface SubCategoryRepository extends JpaRepository<SubCategory, Long> {
 }
